customer_support/requirements.txt
fastapi==0.135.1
groq==1.0.0
pandas==3.0.1
pydantic==2.12.5
python-dotenv==1.2.2
Requests==2.32.5
uvicorn==0.41.0
deepgram-sdk==3.2.1
python-multipart
aiofiles
httpx
